Extract TrelloMember into trello_member.dart
