+++ lib/src/sdk/members/members.dart
@@ -1,3 +1,4 @@
 export 'member_board_filter.dart';
 export 'member_client.dart';
 export 'member_models.dart';
+export 'trello_member.dart';
